Finalize typescript conversion and add more advanced collision detection/handling
Change all files to use idiomatic typescript and have .ts extensions, as well as install
toolchain to compile typescript to javascript. To compile typescript, which must be done
after every change to the code in order to see the results reflected in the website, run
the command "npm run tsc" in the root directory of the project.

Collision detection with bricks should now work, although no bricks have yet been added,
and collisions between the ball and the paddle now take into account where on the paddle
the ball hits in order to change the direction the ball bounces off, as in most classic
brick breaker games. The ball's collision loop also now takes a delta time value, in
preparation for changing the structure of the main loop to have a framerate-independent
logic timestep for smoother animation and handling.