Add brick rendering, update collision handling, and add test brick
Bricks will now render on the screen with a color assigned according to
remaining strength, and will disappear once their strength reaches 0.
Bricks with a strength of 0 are not removed from the DOM, but are hidden
with css and ignored for collision handling. A test brick has also been
added to demonstrate functionality.