Allow placing and removing bricks in level editor
Clicking now places and removes bricks in the level editor depending
on whether or not a brick is already present at the specified
location. In order to correctly keep track of added and removed
bricks for the drawing code, the array of bricks in index.ts has been
changed to a Map of integer id's to bricks instead. This will also
allow actual removal of destroyed bricks during gameplay instead of
just hiding them with css, although the code for that is not yet
implemented.